LayoutTests/fast/frames/seamless/seamless-sandbox-flag-expected.txt

Test that sandbox disables nested seamless iframes correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.seamless is true
PASS nestedFrame.seamless is true
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "150px"
